Created CreditCardValidation.c
CREDIT CARD VALIDATION SYSTEM -- This program is based on LUHN'S ALGORITHM(created in 1970 at IBM). This system figures out if your credit card or debit for the matter is valid or not, if valid it also presents to you the payment network company of the CARD(VISA, AMEX, and MASTERCARD) automatically.